+++ docs/gitbook/dotnet/changelog.md
@@ -1,4 +1,6 @@
-# Changelog
+# 1.0.0 (2026-08-05)
 
-The .NET (`BullMQ` NuGet package) changelog is generated automatically by the
-release pipeline. Entries will appear here after the first `[dotnet]` release.
+
+### Features
+
+* initial implementation of the dotnet port [dotnet] ([#4457](https://github.com/taskforcesh/bullmq/issues/4457)) ([9aab10d](https://github.com/taskforcesh/bullmq/commit/9aab10daa3ce0242055c0e334fac44b12c803243))